fixed check error
