+++ frontend/altinn-support-dashboard.client/src/components/Notification/NIN-search/NotificationShipmentCard.tsx
@@ -1,7 +1,7 @@
 import { Alert, Card, Paragraph, Table } from "@digdir/designsystemet-react";
 import { NotificationShipmentResponse } from "../../../models/notificationModels"
 import styles from "./NotificationShipmentCard.module.css"
-import { colorMap } from "../NotificationStatusCode";
+import { colorMap } from "../notificationColorMap";
 
 
 type NotificationShipemntCardProps = {
